SBERDOMA-575 Extract common component
