+++ apps/condo/domains/contact/hooks/useTableColumns.tsx
@@ -1,43 +1,9 @@
-import { Input, Space } from 'antd'
-import { Button } from '@condo/domains/common/components/Button'
 import { FilterValue } from 'antd/es/table/interface'
 import get from 'lodash/get'
 import { useIntl } from '@core/next/intl'
 import React, { useMemo } from 'react'
-import { FilterFilled } from '@ant-design/icons'
-import { colors } from '@condo/domains/common/constants/style'
 import { createSorterMap, IFilters } from '../utils/helpers'
-
-const getFilterIcon = filtered => <FilterFilled style={{ color: filtered ? colors.sberPrimary[5] : undefined }} />
-
-interface IFilterContainerProps {
-    clearFilters: () => void
-    showClearButton?: boolean
-}
-// TODO (SavelevMatthew) 1 common component in common domain?
-const FilterContainer: React.FC<IFilterContainerProps> = (props) => {
-    const intl = useIntl()
-    const ResetLabel = intl.formatMessage({ id: 'filters.Reset' })
-
-    return (
-        <div style={{ padding: 16 }}>
-            <Space size={8} direction={'vertical'} align={'center'}>
-                {props.children}
-                {
-                    props.showClearButton && (
-                        <Button
-                            size={'small'}
-                            onClick={() => props.clearFilters()}
-                            type={'inlineLink'}
-                        >
-                            {ResetLabel}
-                        </Button>
-                    )
-                }
-            </Space>
-        </div>
-    )
-}
+import { getTextFilterDropdown, getFilterIcon } from '@condo/domains/common/components/TableFilter'
 
 const getFilteredValue = (filters: IFilters, key: string | Array<string>): FilterValue => get(filters, key, null)
 
@@ -52,23 +18,7 @@ export const useTableColumns = (sort: Array<string>, filters: IFilters,
 
     const sorterMap = createSorterMap(sort)
 
-    const getFilterDropdown = (columnName: string) => {
-        return ({ setSelectedKeys, selectedKeys, confirm, clearFilters }) => {
-            return (
-                <FilterContainer clearFilters={clearFilters} showClearButton={selectedKeys && selectedKeys.length > 0}>
-                    <Input
-                        placeholder={columnName}
-                        value={selectedKeys}
-                        onChange={e => {
-                            setSelectedKeys(e.target.value)
-                            setFiltersApplied(true)
-                            confirm({ closeDropdown: false })
-                        }}
-                    />
-                </FilterContainer>
-            )
-        }
-    }
+
 
     return useMemo(() => {
         return [
@@ -80,7 +30,7 @@ export const useTableColumns = (sort: Array<string>, filters: IFilters,
                 key: 'name',
                 sorter: true,
                 width: '25%',
-                filterDropdown: getFilterDropdown(NameMessage),
+                filterDropdown: getTextFilterDropdown(NameMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
             {
@@ -92,7 +42,7 @@ export const useTableColumns = (sort: Array<string>, filters: IFilters,
                 key: 'address',
                 sorter: false,
                 width: '35%',
-                filterDropdown: getFilterDropdown(AddressMessage),
+                filterDropdown: getTextFilterDropdown(AddressMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
             {
@@ -103,7 +53,7 @@ export const useTableColumns = (sort: Array<string>, filters: IFilters,
                 key: 'phone',
                 sorter: true,
                 width: '20%',
-                filterDropdown: getFilterDropdown(PhoneMessage),
+                filterDropdown: getTextFilterDropdown(PhoneMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
             {
@@ -115,7 +65,7 @@ export const useTableColumns = (sort: Array<string>, filters: IFilters,
                 key: 'email',
                 sorter: true,
                 width: '20%',
-                filterDropdown: getFilterDropdown(EmailMessage),
+                filterDropdown: getTextFilterDropdown(EmailMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
         ]
