+++ apps/condo/domains/organization/hooks/useTableColumns.tsx
@@ -1,45 +1,12 @@
 import { identity } from 'lodash/util'
-import { Input, Space, Checkbox } from 'antd'
-import { Button } from '@condo/domains/common/components/Button'
+import { Checkbox } from 'antd'
 import { FilterValue } from 'antd/es/table/interface'
 import get from 'lodash/get'
 import { useIntl } from '@core/next/intl'
 import React, { useMemo } from 'react'
-import { FilterFilled } from '@ant-design/icons'
-import { colors } from '@condo/domains/common/constants/style'
 import { createSorterMap, IFilters } from '../utils/helpers'
 import { OrganizationEmployeeRole } from '../utils/clientSchema'
-
-const getFilterIcon = filtered => <FilterFilled style={{ color: filtered ? colors.sberPrimary[5] : undefined }} />
-
-interface IFilterContainerProps {
-    clearFilters: () => void
-    showClearButton?: boolean
-}
-
-const FilterContainer: React.FC<IFilterContainerProps> = (props) => {
-    const intl = useIntl()
-    const ResetLabel = intl.formatMessage({ id: 'filters.Reset' })
-
-    return (
-        <div style={{ padding: 16 }}>
-            <Space size={8} direction={'vertical'} align={'center'}>
-                {props.children}
-                {
-                    props.showClearButton && (
-                        <Button
-                            size={'small'}
-                            onClick={() => props.clearFilters()}
-                            type={'inlineLink'}
-                        >
-                            {ResetLabel}
-                        </Button>
-                    )
-                }
-            </Space>
-        </div>
-    )
-}
+import { getTextFilterDropdown, getFilterIcon, FilterContainer } from '@condo/domains/common/components/TableFilter'
 
 const getFilteredValue = (filters: IFilters, key: string | Array<string>): FilterValue => get(filters, key, null)
 
@@ -54,23 +21,6 @@ export const useTableColumns = (organizationId: string, sort: Array<string>, fil
 
     const sorterMap = createSorterMap(sort)
     const { loading, objs: organizationEmployeeRoles } = OrganizationEmployeeRole.useObjects({ where: { organization: { id: organizationId } } })
-    const getFilterDropdown = (columnName: string) => {
-        return ({ setSelectedKeys, selectedKeys, confirm, clearFilters }) => {
-            return (
-                <FilterContainer clearFilters={clearFilters} showClearButton={selectedKeys && selectedKeys.length > 0}>
-                    <Input
-                        placeholder={columnName}
-                        value={selectedKeys}
-                        onChange={e => {
-                            setSelectedKeys(e.target.value)
-                            setFiltersApplied(true)
-                            confirm({ closeDropdown: false })
-                        }}
-                    />
-                </FilterContainer>
-            )
-        }
-    }
     const columns = useMemo(() => {
         return [
             {
@@ -81,7 +31,7 @@ export const useTableColumns = (organizationId: string, sort: Array<string>, fil
                 key: 'name',
                 sorter: true,
                 width: '40%',
-                filterDropdown: getFilterDropdown(NameMessage),
+                filterDropdown: getTextFilterDropdown(NameMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
             {
@@ -93,7 +43,7 @@ export const useTableColumns = (organizationId: string, sort: Array<string>, fil
                 sorter: true,
                 width: '20%',
                 render: (position) => position ? position : '—',
-                filterDropdown: getFilterDropdown(PositionMessage),
+                filterDropdown: getTextFilterDropdown(PositionMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
             {
@@ -137,7 +87,7 @@ export const useTableColumns = (organizationId: string, sort: Array<string>, fil
                 key: 'phone',
                 sorter: true,
                 width: '20%',
-                filterDropdown: getFilterDropdown(PhoneMessage),
+                filterDropdown: getTextFilterDropdown(PhoneMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
             {
@@ -147,7 +97,7 @@ export const useTableColumns = (organizationId: string, sort: Array<string>, fil
                 filteredValue: getFilteredValue(filters, 'email'),
                 key: 'email',
                 width: '20%',
-                filterDropdown: getFilterDropdown(EmailMessage),
+                filterDropdown: getTextFilterDropdown(EmailMessage, setFiltersApplied),
                 filterIcon: getFilterIcon,
             },
         ]
